+++ go/engine/texture.go
@@ -4,7 +4,7 @@ import "github.com/veandco/go-sdl2/sdl"
 
 type Texture struct {
 	*sdl.Texture
-	ID uint64
+	ID uint32
 	W  int32
 	H  int32
 }
